pyproject.toml
[project]
name = "ocularis"
version = "0.1.0"
description = "Production-grade orchestration layer for autonomous web agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.41.0",
    "playwright>=1.58.0",
    "httpx>=0.28.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "pyyaml>=6.0.2",
    "loguru>=0.7.3",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg>=0.31.0",
    "pgvector>=0.4.0",
    "openai>=2.20.0",
    "pillow>=12.0.0",
    "scikit-image>=0.26.0",
    "python-multipart>=0.0.20",
    "websockets>=16.0",
    "alembic>=1.18.0",
    "ui-tars>=0.4.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "logic", "api", "db"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
